fix send message through interface -1 (NET_INTERFACE_INVALID)
Previously, the CBR simulation would crash if the fading was changed to
Fading.Rician(2.7), due to an attempted access to the invalid NIC -1.

This invalid interface index was returned from RouteZrpNdp.getMacId(),
and this error return code was not checked. It is not clear if this
reveals a deeper bug (there always should be a valid NIC for the
destination address that triggered the bug -- a bug in the business
logic of RouteZrp) or if this was just an unhandled error code.

Handling the error code and dropping the message to be sent seems to
fix the issue.